// cgoml-ui/src/app/features/models/messaging/views/full/OutboundMessageFullView.ts
import { OutboundMessageForm } from 'src/app/features/models/messaging/forms/OutboundMessageForm';
/**
 * auto generated full veiw bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class OutboundMessageFullView {

id: number;
form: OutboundMessageForm;
canUpdate: boolean;
canDelete: boolean;

/* Specific Code Start */
/* Specific Code End */
}